Support .meta/prefixes.txt
